completed the edit and delete functionality

+++ src/components/ExperianceForm.jsx
@@ -7,6 +7,10 @@ import axios from 'axios';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Edit from './Edit';
+import './EducationalForm.css';
+import sty from "./formatedStyle.module.css";
+import Model from './Model';
+
 
 function ExperianceForm() {
     const [status, setShowStatus] = useState(false);
@@ -26,6 +30,7 @@ function ExperianceForm() {
 
 
     const [experiences, setExperiences] = useState([]);
+
 
 
 
@@ -93,7 +98,7 @@ function ExperianceForm() {
                 lastDate: new Date(formsData.lastDate).toISOString()
             };
 
-            fetch(process.env.REACT_APP_SERVER+'/api/preexp', {
+            fetch('http://192.168.2.81:5003/api/preexp', {
                 method: "POST",
                 headers: { "content-type": "application/json" },
                 body: JSON.stringify(payload)
@@ -118,26 +123,46 @@ function ExperianceForm() {
                 });
         }
     };
-    useEffect(() => {
-        const fetchEvents = () => {
-            axios.get(process.env.REACT_APP_SERVER+'5003/api/GetPrevExp/7')
-                .then((resposne) => {
+    const fetchEvents = () => {
+        axios.get('http://192.168.2.81:5003/api/GetPrevExp/7')
+            .then((resposne) => {
 
-                    if (resposne?.status === 200) {
-                        toast(resposne.message);
-                        setExperiences(resposne?.data)
-                        console.log(resposne?.data);
-                    } else {
-                        alert('something went wrong')
-                    }
-                    if(resposne?.status === 405){
-                        toast(resposne.message);
-                    }
-                })
-                .catch((error) => {
-                    return error
-                })
-        }
+                if (resposne?.status === 200) {
+                    toast(resposne.message);
+                    setExperiences(resposne?.data)
+                    console.log(resposne?.data);
+                } else {
+                    alert('something went wrong')
+                }
+                if (resposne?.status === 405) {
+                    toast(resposne.message);
+                }
+            })
+            .catch((error) => {
+                return error
+            })
+    }
+    useEffect(() => {
+        //REMOVE THIS
+        // const fetchEvents = () => {
+        //     axios.get(process.env.REACT_APP_SERVER+'5003/api/GetPrevExp/7')
+        //         .then((resposne) => {
+
+        //             if (resposne?.status === 200) {
+        //                 toast(resposne.message);
+        //                 setExperiences(resposne?.data)
+        //                 console.log(resposne?.data);
+        //             } else {
+        //                 alert('something went wrong')
+        //             }
+        //             if(resposne?.status === 405){
+        //                 toast(resposne.message);
+        //             }
+        //         })
+        //         .catch((error) => {
+        //             return error
+        //         })
+        // }
         fetchEvents();
     }, [formsData])
 
@@ -168,28 +193,53 @@ function ExperianceForm() {
         console.log("handleClosemethod triggered")
         setEditStatus(false);
         fetchEvents();
+    };
+
+    //i am trying to keep whole delete functionality in this area below
+    const [deleteStatus, setDeleteStatus] = useState(false);
+    const [itemToDelete, setItemToDelete] = useState([])
+
+
+    const handleDelete = async (val) => {
+        setDeleteStatus(true);
+        setItemToDelete(val);
+
     }
 
 
+    const confirmDelete = async () => {
+        if (!itemToDelete) return;
 
-    const handleDelete = (id) => {
-
-        const url = `http://192.168.2.81:5003/api/preexp/${id}`;
+        const url = `http://192.168.2.81:5003/api/preexp`;
         try {
-            fetch(url, {
+            await fetch(url, {
                 method: "DELETE",
                 headers: {
                     "content-type": "application/json"
-                }
+                },
+                body: JSON.stringify(itemToDelete)
+            }).then(()=>{
+                toast('Your previous experiance has been successfully deleted');
             })
 
         }
         catch (error) {
             console.log(error);
         }
-      fetchEvents();
-      console.log(fetchEvents)
+        fetchEvents();
+        setItemToDelete(null);
+        setDeleteStatus(false);
+        console.log(fetchEvents)
     }
+
+    const cancelDelete = () => {
+        setDeleteStatus(false);
+        setItemToDelete(null);
+    }
+
+
+
+    //i am trying to keep whole delete functionality in this area above
 
 
     return (
@@ -198,7 +248,7 @@ function ExperianceForm() {
             <form onSubmit={handleSubmit} className='formfull'>
                 <div className='formitem1'>
                     <div className='inneritems'>
-                        <div className='form-group'>
+                        <div className={sty.form_group}>
                             <input
                                 type="text"
                                 id="messi"
@@ -208,10 +258,11 @@ function ExperianceForm() {
                                 onChange={handleChange}
                                 name='companyName'
                                 required
+                            // autoFocus="off"
                             />
                             <label htmlFor="messi"><span className="star">*</span>Company Name</label>
                         </div>
-                        <div className='form-group'>
+                        <div className={sty.form_group}>
                             <input
                                 type="text"
                                 id="messi2"
@@ -224,7 +275,7 @@ function ExperianceForm() {
                             />
                             <label htmlFor="messi2"><span className="star">*</span>Designation</label>
                         </div>
-                        <div className='form-group'>
+                        <div className={sty.form_group}>
                             <input
                                 type="date"
                                 id="messi3"
@@ -238,7 +289,7 @@ function ExperianceForm() {
                             <label htmlFor="messi3"><span className="star">*</span>Start Date</label>
                             {errors.startDateError && <p className='error'>{errors.startDateError}</p>}
                         </div>
-                        <div className='form-group'>
+                        <div className={sty.form_group}>
                             <input
                                 type="date"
                                 id="messi4"
@@ -261,33 +312,55 @@ function ExperianceForm() {
 
 
 
-                    <div className='inneritems'>
-                        <ul style={{ width: '-webkit-fill-available' }} className='topside'>
-                            {experiences.map((exp, index) => (
-                                <li key={index} className='listed'>
-                                    <p className='index'>{index + 1}</p>
-                                    <p className='form-group' id="ff"> {exp.companyName}</p>
-                                    <p className='form-group' id="ff">{exp.designation}</p>
-                                    <p className='form-group' id="ff">{formatDate(exp.startDate)}</p>
-                                    <p className='form-group' id="ff">{formatDate(exp.lastDate)}</p>
-                                    <p className='form-group' id='edit-icon' onClick={() => handleEdit(exp.id)}>
-                                        <svg width="30" height="37" viewBox="0 0 42 50" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <path d="M31.1576 0.571289L25.9647 6.69629L36.3505 18.9463L41.5435 12.8213L31.1576 0.571289ZM20.7717 12.8213L0 37.3213V49.5713H10.3859L31.1576 25.0713L20.7717 12.8213Z" fill="black" />
-                                        </svg>
-                                    </p>
-                                    <p id='delete-icon' onClick={handleDelete(exp.id)}>
-                                        <svg width="30" height="37" viewBox="0 0 41 50" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <path d="M7.6875 49.5713C6.27812 49.5713 5.07204 49.0386 4.06925 47.9733C3.06646 46.908 2.56421 45.6259 2.5625 44.1268V8.73796H0V3.29351H12.8125V0.571289H28.1875V3.29351H41V8.73796H38.4375V44.1268C38.4375 45.6241 37.9361 46.9062 36.9333 47.9733C35.9305 49.0405 34.7236 49.5731 33.3125 49.5713H7.6875ZM33.3125 8.73796H7.6875V44.1268H33.3125V8.73796ZM12.8125 38.6824H17.9375V14.1824H12.8125V38.6824ZM23.0625 38.6824H28.1875V14.1824H23.0625V38.6824Z" fill="black" />
-                                        </svg>
-                                    </p>
-                                </li>
-                            ))}
-                        </ul>
-                    </div>
+
+                    {experiences.length > 0 ? (
+                        <table className='tabl'>
+                            <thead className='tablehead'>
+                                <tr className='top-heading'>
+                                    <th>CompanyName</th>
+                                    <th>Designation</th>
+                                    <th>StartDate</th>
+                                    <th>EndDate</th>
+                                    <th>Actions</th>
+                                    <th></th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {experiences.reverse().map((val, index) => (
+                                    <tr className='decrease-me-there' key={index}>
+                                        <td className='font-correction'>{val.companyName}</td>
+                                        <td className='font-correction'>{val.designation}</td>
+                                        <td className='font-correction'>{formatDate(val.startDate)}</td>
+                                        <td className='font-correction'>{formatDate(val.lastDate)}</td>
+
+                                        <td>
+                                            <p className='form-group' id='edit-icon' onClick={() => handleEdit(val.id)}>
+                                                <svg width="22" height="30" viewBox="0 0 42 50" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                                    <path d="M31.1576 0.571289L25.9647 6.69629L36.3505 18.9463L41.5435 12.8213L31.1576 0.571289ZM20.7717 12.8213L0 37.3213V49.5713H10.3859L31.1576 25.0713L20.7717 12.8213Z" fill="black" />
+                                                </svg>
+                                            </p>
+                                        </td>
+                                        <td>
+                                            <p className='delete-button' id='delete-icon' onClick={() => handleDelete(val)}>
+                                                <svg width="22" height="30" viewBox="0 0 41 50" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                                    <path d="M7.6875 49.5713C6.27812 49.5713 5.07204 49.0386 4.06925 47.9733C3.06646 46.908 2.56421 45.6259 2.5625 44.1268V8.73796H0V3.29351H12.8125V0.571289H28.1875V3.29351H41V8.73796H38.4375V44.1268C38.4375 45.6241 37.9361 46.9062 36.9333 47.9733C35.9305 49.0405 34.7236 49.5731 33.3125 49.5713H7.6875ZM33.3125 8.73796H7.6875V44.1268H33.3125V8.73796ZM12.8125 38.6824H17.9375V14.1824H12.8125V38.6824ZM23.0625 38.6824H28.1875V14.1824H23.0625V38.6824Z" fill="black" />
+                                                </svg>
+                                            </p>
+                                        </td>
+                                    </tr>
+                                ))}
+                            </tbody>
+                        </table>
+                    ) : (
+                        <p className="centertext">No previous experiences details added</p>
+                    )}
                 </div>
-                {status && <PopeUp submitconfirm={handleConfirm} submitcancel={handleCancel} />}
-                {editstatus && <Edit status={editstatus} close={handleClose} ide={identity} />}
+
             </form>
+
+            {status && <PopeUp submitconfirm={handleConfirm} submitcancel={handleCancel} />}
+            {editstatus && <Edit status={editstatus} close={handleClose} ide={identity} />}
+            {deleteStatus && <Model message='Are you sure you want to Delete Previous Experiance!!' confirmHandler={confirmDelete} cancelHandler={cancelDelete} />}
         </div>
     );
 }
